requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0
PyJWT==2.8.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
openai==1.3.7
reportlab==4.0.7
matplotlib==3.8.2
numpy==1.25.2
Pillow==10.1.0 